+++ scripts/configs/networks/optimism.ts
@@ -1,6 +1,6 @@
 import {Hex} from 'viem';
 import {ChainId} from '@bgd-labs/toolbox';
-import {NetworkAddresses} from '../types';
+import {NetworkAddresses} from 'scripts/configs/types';
 
 export const optimismAddresses: NetworkAddresses<{
   wstETH_stETH_AGGREGATOR: Hex;