+++ scripts/configs/networks/sonic.ts
@@ -1,5 +1,5 @@
 import {ChainId} from '@bgd-labs/toolbox';
-import {NetworkAddresses} from '../types';
+import {NetworkAddresses} from 'scripts/configs/types';
 
 export const sonicAddresses: NetworkAddresses = {
   name: 'Sonic',